Move mocks for gtest_paralell_tests.py to another file.
